[server] IPv6 socket support, and binding to multiple sockets
Bind to each address returned by getaddrinfo(), in the manner
(PF_UNSPEC) used to bind to both IPv4 and IPv6 wildcard addresses on
the specified port.

This also lays the groundwork for specifying multiple listen addresses
in the config file, in particular the common case of one SSL port
and one non-SSL port.