CLD: Sanitize memory handling of DB4
This is necessary because DB_DBT_MALLOC cannot be set conditionally
for a database with DB_THREAD. If we try, database access bombs with
a message. Although we know that nothing would be returned, DB does not.
DB_THREAD, in turn, is required if we ever want to implement distributed
database. This is a completely reasonable behaviour of the DB, because
the default memory management is decidedly thread-unsafe. Better correct
it now than later.

Since we're on it, using DB_DBT_MALLOC is unnecessary in many cases.
We use DB_DBT_USERMEM and auto structs to avoid tracking and free(),
where appropriate.

Also, fix the type confusion. We passed "true" as a flag. Fortunately,
0x1 is an invalid DB flag and so we received an error message instead
of possibly hard to detect malfunction.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>