CLD: Add the root inode
It's easier to add the root node from the daemon than write whole cldbadm.
And it may be advantageous to have the daemon to auto-start over an
empty directory. So, although cldbadm is inevitable, create the root
in the daemon for now.

This also includes dealing with a freshly created directory with
no directory data, which is needed for all directories, not just root.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>