libcldc: Expire unopened sessions
This is needed to make sessions to establish if applications can retry.

First of all, we were not running timers for retry messages, so any
little failure or lost packet would cause sessions to fail to open.
Fix this by re-establishing the timer when we resend messages.

Secondly, session expiration was all broken. We used a global
cldc_current_time, set when packets arrived. But if packets NOT
arrived at all (e.g. if CLD was not running), cldc_current_time
was zero we expired sessions immediately. The fix is to use
actual time in timer handlers. We keep the optimization, but
to make it harder to misuse, we propogate the time as argument
instead of a global variable. This way, if a routine is not
called from a receive handler, it has no access to the cached time.

But wait, there's more: we forgot to set the expiration time
when we created the session. The fix is obvious.

With these fixes in place, even if session was not opened, the
application will receive an event after 2 minutes.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>