cld,libcldc: add SIDFMT, SIDARG
This patch adds a uniform format in case someone wants to print
a Session ID. With it, it's easier to scan logfiles (hit '*' in vi).

We can use lowercase hex, or decimal, it's not important. The important
part is to print the same thing in all places.

This patch also makes sid2llu available to clients using libcldc.
The way we do it is a little convoluted, and we also abuse cld_msg.h
as a common header. The hope is that the future evolution of the library
will force some presently unforseeable cleanup and refactoring.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>