libcldc: Accept terminating segment
Server sends us a terminating zero length segment, but we expire the
stream when all data is received. Upon receipt of the terminator,
we return error and tests become very confused.

We probably should not remove the terminator from the protocol,
in case we have meta-only transfers. It seems more reliable this way.
The put method is similarly terminated, BTW. So, the fix is to wait
until the terminator and only expire the stream then.

Also, we forgot to increment the last segment number, causing erroneous
mismatches.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>