libcldc: Add buf, len to get API
I don't see any way for the application actually access the data that
cldc received and stored in the stream. The application cannot not know
how to access the stream into which the copts are embedded, the stream
is a concept internal to cldc.

Add the pointer that can be examined. Bloat and punishing all other
calls and callbacks, but it's expedient. Maybe install a different
callback for get request later?

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>