cld: Load old sessions on boot
If we do not load sessions, old locks remain stuck forever and no client
can continue, because the locks cannot be cleared.

I remember now what the problem was with this patch: it did not reset
any timers, so it there were good chances that reloaded sessions would
expire right away (which suited me because the main problem was with
the stuck locks). Timers can be fixed, but I don't want to disrupt
anything for this patch group.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>