cldc: fix crash in get_end_cb
If cmo_get returns an error, no inode is returned. However, use the
inode layout, and then write all over the stack with wrong o->ino_len
(it overlays with sequence number in such case).

Solution: only interpret inode if success was returned. It's analoguous
to the code in open_end_cb.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>