cld: don't rely on unrelated struct layouts
The code is correct for now, but the trick with structs defined in
different headers needing to have the same fields looks terribly dangerous
to me, and I would sleep better if the fix were applied. It was posted
previously along with a patch that removed the comment warning about
the field layout across structures, and that got applied.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>