chunkd: make config.c similar to tabled's
This patch changes config.c to be more similar to its equivalent in tabled.
In particular, static variables are being eradicated in case we create a
common library. Unfortunately, there's still a lot of application-specific
code embedded into either config.c, and we don't have a good place for
the shared code, so this work is not complete. At least copy-paste is
safer now.

A side effect is that now -P on command line overrides the configuration,
instead of backwards as before, which violated the KISS principle.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>