cld: redirect logging
When we build in Koji, the syslog is not available and we cannot see
what daemons print that we spawn for "make check". Thus debugging
of build problems is impeded.

This patch allows to redirect the log of daemon. We already have
a similar facility for libcldc because it's linked from both interactive
programs and daemons such as tabled, but this covers the rest of CLD.

For simplicity and transparency, we don't play any games with redefining
syslog(3) or function pointers.

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>