chunkd: write our contact information into CLD
Add our listening information into the parameter file in CLD.

For now we only do that, and do not try to reorganize the CLD
namespace that Chunk uses, use full path as a parameter, and other
planned things. We do switch to asprintf though, it makes the
code more robust (no more counting of elements by hand).

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>