libcldc: transition to applog, phase 1
This patch transitions from printf-like to syslog-like API for
libcldc diagnostics. In the future, this allows to drop app_log()
from chunkd (tabled didn't even grow it).

There's a small complication: one stdarg function cannot call
another, so we cannot just insert some shims... Or actually, we can.
The trick is to use printf into a local buffer, and then pass that down.
All of that is strictly temporary. If I'm not missing something, once
applications switch over, we'll just kill all of it and return to
what it was, only with syslog-like functions. I didn't even adjust
whitespace of ->app_log arguments as a pledge that it has to change
again anyway.

Long story short, this should be safe to build in Koji without
resorting to chain builds.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>