cldcli: permit (much) longer messages
A thought just occured to me that it's dumb to expand the message
itself when you can expand the format. This way all the potentially
long things like CLD pathnames do not count against your stack
and there's essentially no chance to get any messages truncated.

The first snprintf is actually a strncat, but it's more concise.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>