cld: return more correct codes
When doing "ls /foo" in cldcli currently, server reports "invalid inode
number". But in fact it's a name that's incorrect, all numbers are fine.
It seems we need to return "invalid name" afer we looked up by name,
and "invalid number" after we looked up by number (this is ok now and
not changed by the patch).

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>