cld: implement dynamic port
This patch is needed to build on Fedora. The buildsystem can run two
builds on the same system in the same time (for example, one for 32-bit
buildroot, and another for 64-bit buildroot). Since we include
"make check" step, port numbers used by simultaneous builds conflict,
causing the build to fail.

Keeping a patch for this in Fedora is additional labor placed on
maintainer, and there's a remote possibility that it may be useful
for someone else, so let's include this feature.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>