chunkd: implement auto ports
This patch helps us to build inside Fedora build system where port
conflicts from simultaneous builds are an issue. Most of the code is
taken from a similar patch in CLD, although there are differences.
In particular, we use the configuration file instead of command line.

The whole deal is left underdocumented.

Also, this patch changes the invocation of cld in start-daemon,
and therefore requires changes in the installed cld. Thus we have
a one-way flag day.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>