chunkd: use "auto" keyword
The current implementation of automatic listening ports was configured
implicitly by adding <PortFile> clause. This turned out to be a silly
idea. A special keyword for <Port> is more convenient. Also, we do not
need to use /dev/null in <PortFile> anymore.

The configuration is not strictly speaking compatible, but since
nobody should be using auto ports in production anyway, impact on
installations should be nil. But this must be built after the support
for "-p auto" is added to CLD.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>