cld: update storage backend to store based on opaque binary key + key length
Change internal structure to utilize an opaque binary key, rather than the
existing probably-nul-terminated, fixed length C hexidecimal string key.

Storage backend now names files based on the SHA256 of the key.

The key, in its entirety, is written to each object's header in the
filesystem, to assist in recovery.

This work prepares the server for an API and protocol change permitting
arbitrary length binary keys.

Signed-off-by: Jeff Garzik <jgarzik@redhat.com>