CLD: drop dependency on libevent from libcldc
Let's remove the dependency on libevent completely. The core CLD daemon
did not use libevent for some time, but getting rid of it entirely required
an implementation of some kind of common main loop for tests, which is
what this patch does.

Since we were refactoring, we add a common header and library for
tests, same as Chunk has.

We use the same timers that the main daemon uses (libtimer), they are
general enough for other applications such as tests.

This patch introduces an incompatibility, so Chunk and tabled must
be rebuilt. This is because the layout of struct cldc_udp changes.

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>