CLD: export a common timer implementation
Although we share the libtimer.a among the parts that can link it
statically, it turned out impossible to use from inside of libcldc
as it was. So, this patch makes it available to library code by
linking it in the same way we use libcommon.c already.

But if we do that, timer_add and friends become visible to all
users of libcldc, so rename them with cld_ prefix in order to
reduce namespace pollution.

Also, make the list of timers anchored outside of the library.
Not the we know of anyone who needs two in one program, but
it costs us almost nothing and is safer for the future.

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>