CLD: make cldcli to use common timer implementation
Good diffstat balance: 57 deletions over 8 insertions.

Actually, this code did not need all the renames and changes in the
way libtimer.c was built, because cldcli was built inside the CLD tree
and could thus use the old libtimer.a. But using the new libtimer is
more honest, to have code copy-pasted elsewhere.

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>