Rename cells to groups
The term "cell" in context of tabled and chunkd conflicts with the idea
of a CLD cell. So, rename it to "group".

Notice that we aren't introducing configuring of actual CLD paths yet.
Coding that in the old CLD API is too hard, so it will come after ncld.
Stay tuned.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>