kill CLD_MAX_PKT_MSG, add CLD_MAX_PAYLOAD_SZ
Get rid of CLD_MAX_PKT_MSG. It only existed so that we could use
static arrays in a few places.

Create CLD_MAX_PAYLOAD_SZ to represent the maximum size of a message
that the API user can GET or PUT. Reducing this constant could break
users who relied on the old maximum data size, so we should try not
to do it often.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>