libcldc: Introduce the "New CLD" API
The "traditional" CLD API is too difficult to program. In particular,
switching from a rigid "Group" policy to arbitrary paths in existing
clients in Chunk and tabled turned out to be next to impossible.
The issue is due to the fundamental nature of the API as based on very
fine-grained events.

So, in the interests of clients, introduce the "new" API. Its basic
definition is in <ncld.h>. It presents a filesystem-like interface,
extended with event callbacks for coarse events, such as an end of
CLD session.

The patch converts all in-tree clients from cldc_xxx to ncld_xxx and
discards some of code that became unused (test/util.c).

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>