libtimer: fix regression in unified cld_timer
The common timer implementation had a bug whereas it overwrote the list
of timers when a timer callback added its own timer -- which just about
every one of them does. The most visible result was that sessions stopped
expiring correctly and left dangling file locks. Also, every new session
failed after 2 minutes 35 seconds.

The fix is not to keep a local copy of the list head across invocations
of the callback.

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>