CLD: End-to-end verbosity v3
It has been observed that it's not possible to enable the session-level
verbose dump with -v switch of cldcli anymore (post XDR and ncld).

The main issue is that it is impossible to set the session verbosity by
an application control. Secondary issue is that if application attempts
to change the verbosity level, it cannot do so for newly-created sessions,
because the session structure is allocated when session is created.

Contributing to the confusion is the desire not to enable excessive
verbosity in the CLD sessions whenever debugging is requested.

To fix these issues, the patch does the following:

 - Drops all commented-out inserts that assign additional verbosity
   from libraries. Permits top-level applications to specify verbosity
   (what cldcli -v was intended to do).

 - Splits per-packet verbosity away from debugging messages.

 - Establishes -D 2 to mean "CLD protocol verbosity" across applications
   and the daemon (this replaces the "old" -v). The existing -D 1 now
   controls the program debugging only, and there is no bit mask.

Coincidentially this fixes the crash whenever there's a resolution
failure for an SRV record.

Signed-Off-By: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>