libcldc: fix hang in ncld_sess_close
The use model of ncld made one thing obvious: the cldc_close has
two functions: it disposes of the handle in the client memory
(in theory -- in practice we do not free those until session
terminates), and also talks to the server about that. Our operations
have no own timers, so if session goes down, and someone tries to
close handles of it, the operation hangs forever. The ncld_close
could look at the status of cldc session, but it is a bit of a
layering violation. So, we make cldc_close to verify that the
session is not expired before proceeding.

It would probably be more consistent to make every cldc operation
do that, but the situation arises in practice only during closing
something without knowing if it's dead or alive.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>