chunkd: Protect the invariants
I am unable to ensure that the invariants such as having something in
write_q upon restart are not violated, ever. So, let's rearrange the
code a little bit to make sure.

[Before sending I noticed that this patch included a piece of other,
unrelated patch. However, this passed tests on its own (because
cli_wr_set_poll never returns an error). So in the interests of
not screwing up further, sending as-is.]

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>