chunkd: eliminate double-writing of per-file object metadata
Previously, a PUT would
	<write metadata header>
	<write data>
	<seek to beginning>
	<write metadata header>

Eliminate the double-writing of the initial obj header by simply
seek'ing to the data area immediately after opening the file.
After the object is transferred and metadata may be filled in, write
metadata fully at that time (immediately prior to close).

This reduces the amount of data written, should be more efficient
for small objects, and eases future changes to object metadata area.

Signed-off-by: Jeff Garzik <jgarzik@redhat.com>